rename to CalculationResults to calculation_results
