+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/grid_loss_aggregators.py
@@ -15,7 +15,7 @@ import pyspark.sql.functions as f
 from pyspark.sql import DataFrame
 
 import package.calculation.energy.aggregators.transformations as t
-from package.calculation.energy.energy_results import EnergyResults
+from package.calculation.energy.data_structures.energy_results import EnergyResults
 from package.calculation.preparation.data_structures.grid_loss_responsible import (
     GridLossResponsible,
 )
