+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/grouping_aggregators.py
@@ -15,7 +15,7 @@
 from package.calculation.energy.aggregators.transformations.aggregate_sum_and_quality import (
     aggregate_sum_quantity_and_qualities,
 )
-from package.calculation.energy.energy_results import EnergyResults
+from package.calculation.energy.data_structures.energy_results import EnergyResults
 from package.constants import Colname
