+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/metering_point_time_series_aggregators.py
@@ -14,7 +14,7 @@
 
 import pyspark.sql.functions as f
 
-from package.calculation.energy.energy_results import EnergyResults
+from package.calculation.energy.data_structures.energy_results import EnergyResults
 from package.calculation.energy.aggregators.transformations.aggregate_sum_and_quality import (
     aggregate_quantity_and_quality,
 )
