+++ source/databricks/calculation_engine/package/calculation/energy/data_structures/energy_results.py
