+++ source/databricks/calculation_engine/package/calculation/wholesale/get_wholesale_metering_point_time_series.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import pyspark.sql.functions as f
 
-from package.calculation.energy.energy_results import EnergyResults
+from package.calculation.energy.data_structures.energy_results import EnergyResults
 from package.calculation.preparation.data_structures.prepared_metering_point_time_series import (
     PreparedMeteringPointTimeSeries,
 )
